Refactor code for improved clarity and consistency; update type casting in Racer.c, enhance CSV parsing in PaletteLayer.c, and adjust struct definitions in crt.c and crt.h.

+++ samples/Racer.c
@@ -99,7 +99,7 @@ int main(void) {
     TasksActors(time);
 
     /* render to window */
-    TLN_DrawFrame(time);
+    TLN_DrawFrame((int)time);
 
     frame++;
   }
